[feedback] MapView 계산 로직 분리 (mapCalculator, 함수 모음)
